- ios and android rn version and netmera upgrade were developed
