+++ android/app/src/main/java/com/netmerareactnativeexample/ReactNativeFlipper.java
