--- android/app/src/main/java/com/netmeratsexample/MainActivity.java
@@ -1,34 +1,0 @@
-package com.netmeratsexample;
-
-import com.facebook.react.ReactActivity;
-import com.facebook.react.ReactActivityDelegate;
-import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint;
-import com.facebook.react.defaults.DefaultReactActivityDelegate;
-
-public class MainActivity extends ReactActivity {
-
-  /**
-   * Returns the name of the main component registered from JavaScript. This is used to schedule
-   * rendering of the component.
-   */
-  @Override
-  protected String getMainComponentName() {
-    return "NetmeraTSExample";
-  }
-
-   /**
-     * Returns the instance of the {@link ReactActivityDelegate}. Here we use a util class {@link
-     * DefaultReactActivityDelegate} which allows you to easily enable Fabric and Concurrent React
-     * (aka React 18) with two boolean flags.
-     */
-  @Override
-   @Override
-    protected ReactActivityDelegate createReactActivityDelegate() {
-      return new DefaultReactActivityDelegate(
-          this,
-          getMainComponentName(),
-          // If you opted-in for the New Architecture, we enable the Fabric Renderer.
-          DefaultNewArchitectureEntryPoint.getFabricEnabled());
-    }
-  }
-}
