+++ ios/.xcode.env
@@ -1,6 +1,6 @@
-# This `.xcode.env` file is versioned and is used to source the environment
+# This `.xcode..env` file is versioned and is used to source the environment
 # used when running script phases inside Xcode.
-# To customize your local environment, you can create an `.xcode.env.local`
+# To customize your local environment, you can create an `.xcode..env.local`
 # file that is not versioned.
 # NODE_BINARY variable contains the PATH to the node executable.
 #
