+++ ios/GoogleService-Info.plist
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CLIENT_ID</key>
+	<string>850720650930-9iguskhrovohdlcv8927e1op2noh5has.apps.googleusercontent.com</string>
+	<key>REVERSED_CLIENT_ID</key>
+	<string>com.googleusercontent.apps.850720650930-9iguskhrovohdlcv8927e1op2noh5has</string>
+	<key>API_KEY</key>
+	<string>AIzaSyCxrxAgFKMZ0BwtTIAPQg_akjkD6LW1StA</string>
+	<key>GCM_SENDER_ID</key>
+	<string>850720650930</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.netmerareactnativeexample</string>
+	<key>PROJECT_ID</key>
+	<string>netmeraexamples</string>
+	<key>STORAGE_BUCKET</key>
+	<string>netmeraexamples.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:850720650930:ios:1e40563e97619000657bab</string>
+</dict>
+</plist>
