+++ ios/NetmeraTSExample/AppDelegate.h
@@ -1,12 +1,11 @@
 #import <RCTAppDelegate.h>
+#import "RNCConfig.h"
 #import <UIKit/UIKit.h>
 #import <Netmera/Netmera.h>
 #import <NetmeraCore/NetmeraPushObject.h>
 #import <UserNotifications/UserNotifications.h>
-#import <React/RCTLinkingManager.h>
-#import "RNCConfig.h"
 
 
-@interface AppDelegate : RCTAppDelegate
+@interface AppDelegate : RCTAppDelegate <UNUserNotificationCenterDelegate,NetmeraPushDelegate>
 
 @end
