+++ package.json
@@ -11,19 +11,23 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-native-async-storage/async-storage": "^1.21.0",
     "@react-native-picker/picker": "^2.3.1",
     "@react-navigation/bottom-tabs": "^6.2.0",
     "@react-navigation/native": "^6.0.8",
+    "@react-navigation/native-stack": "^6.9.17",
     "install": "^0.13.0",
     "npm": "^7.24.2",
     "react": "18.2.0",
     "react-hook-form": "^7.28.0",
     "react-native": "0.72.8",
-    "react-native-config": "^1.4.5",
-    "react-native-netmera": "^1.5.0",
+    "react-native-config": "^1.5.1",
+    "react-native-date-picker": "^4.3.5",
+    "react-native-netmera": "^1.6.3",
     "react-native-safe-area-context": "^4.1.3",
     "react-native-screens": "^3.13.1",
-    "react-native-select-dropdown": "^1.8.0"
+    "react-native-select-dropdown": "^1.8.0",
+    "react-native-toast-message": "^2.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
